check for ti_error

+++ helper.sh
@@ -73,9 +73,11 @@ build_docker() {
 test_docker() {
   test_docker_variables
   Rscript example.R /tmp/example.h5
-  sudo docker run -v /tmp:/mnt $REPO:v$VERSION --dataset /mnt/example.h5 --output /mnt/output.h5
-  Rscript -e 'names(dynwrap::calculate_trajectory_dimred(dynutils::read_h5("/tmp/output.h5")))'
-  sudo rm /tmp/example.h5 /tmp/output.h5
+  if [ "$REPO" != "dynverse/ti_error" ]; then 
+    sudo docker run -v /tmp:/mnt $REPO:v$VERSION --dataset /mnt/example.h5 --output /mnt/output.h5
+    Rscript -e 'names(dynwrap::calculate_trajectory_dimred(dynutils::read_h5("/tmp/output.h5")))'
+    sudo rm /tmp/example.h5 /tmp/output.h5
+  fi
 }
 
 push_docker() {
